+++ example/drum_machine.f90
@@ -1,10 +1,10 @@
 ! Forsynth: a multitracks stereo sound synthesis project
 ! License GPL-3.0-or-later
 ! Vincent Magnin, 2024-04-25
-! Last modifications: 2024-04-25
+! Last modifications: 2024-05-13
 
 program drum_machine
-    use forsynth, only: dp, create_WAV_file, &
+    use forsynth, only: dp, create_WAV_file, mix_tracks, &
                       & finalize_WAV_file, clear_tracks, file_t
     use signals, only: add_karplus_strong_drum, add_karplus_strong_drum_stretched
 
@@ -26,6 +26,7 @@ program drum_machine
     end do
 
     print *, "Final mix..."
+    call mix_tracks()
     call finalize_WAV_file()
 
 end program drum_machine